Add functionality to create and display new projects
A method `addProject()` was added to the `Profile` class for creating new projects. The `ProjectSelectBar` class was updated to include a button for adding new projects to a profile. When a user clicks the "+" button, a dialog is opened for the user to input the name of the new project. Upon confirmation, the new project is created, added to the profile, and displayed on the bar.