Add tooltip for budget items and enhance detail view
Updated the app to provide a tooltip for each budget item, displaying the item's name, quantity, and formatted cost. Enhanced the detail view to not only display a list of logs but also allow the user to select a log from the list and view or amend details for that log. Minor GUI improvements were also made in both views..