Rename BudgetView and add item description tooltip
The BudgetView file has been renamed to BudgetView.java. This commit also implements a new feature to the BudgetView where a tooltip pops up showing the description of an item when said item is double-clicked. Code documentation has been added to the BudgetView and minor changes have been made to its structure.